wallet: Deleted some false texts and moved some files for preventing licence problems
