+++ decentra_network/accounts/get_balance.py
@@ -8,7 +8,7 @@ import sqlite3
 
 from decentra_network.accounts.get_accounts import GetAccounts
 from decentra_network.blockchain.block.get_block import GetBlock
-from decentra_network.wallet.ellipticcurve.wallet_import import Address
+from decentra_network.wallet.wallet_import import Address
 
 
 def GetBalance(user,
