+++ decentra_network/accounts/get_sequence_number.py
@@ -5,7 +5,7 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at https://mozilla.org/MPL/2.0/.
 from decentra_network.accounts.get_accounts import GetAccounts
-from decentra_network.wallet.ellipticcurve.wallet_import import Address
+from decentra_network.wallet.wallet_import import Address
 
 
 def GetSequanceNumber(user, account_list=None):
