+++ decentra_network/gui_lib/libs/baseclass/wallet_screen.py
@@ -15,11 +15,11 @@ from decentra_network.blockchain.block.get_block import GetBlock
 from decentra_network.lib.qr import qr
 from decentra_network.lib.settings_system import change_wallet
 from decentra_network.lib.settings_system import the_settings
-from decentra_network.wallet.ellipticcurve.get_saved_wallet import \
+from decentra_network.wallet.get_saved_wallet import \
     get_saved_wallet
-from decentra_network.wallet.ellipticcurve.wallet_create import wallet_create
-from decentra_network.wallet.ellipticcurve.wallet_delete import wallet_delete
-from decentra_network.wallet.ellipticcurve.wallet_import import wallet_import
+from decentra_network.wallet.wallet_create import wallet_create
+from decentra_network.wallet.wallet_delete import wallet_delete
+from decentra_network.wallet.wallet_import import wallet_import
 
 
 class WalletScreen(MDScreen):
