+++ decentra_network/wallet/wallet_selector.py
@@ -6,7 +6,7 @@
 # file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
 
-from decentra_network.wallet.ellipticcurve.get_saved_wallet import get_saved_wallet 
+from decentra_network.wallet.get_saved_wallet import get_saved_wallet 
 
 from decentra_network.wallet.print_wallets import print_wallets
