+++ tests/unit_tests/test_api.py
@@ -56,12 +56,12 @@ from decentra_network.transactions.my_transactions.save_to_my_transaction import
 from decentra_network.transactions.pending.delete_pending import DeletePending
 from decentra_network.transactions.pending.get_pending import GetPendingLen
 from decentra_network.transactions.transaction import Transaction
-from decentra_network.wallet.ellipticcurve.get_saved_wallet import \
+from decentra_network.wallet.get_saved_wallet import \
     get_saved_wallet
-from decentra_network.wallet.ellipticcurve.save_wallet_list import \
+from decentra_network.wallet.save_wallet_list import \
     save_wallet_list
-from decentra_network.wallet.ellipticcurve.wallet_create import wallet_create
-from decentra_network.wallet.ellipticcurve.wallet_import import (Address,
+from decentra_network.wallet.wallet_create import wallet_create
+from decentra_network.wallet.wallet_import import (Address,
                                                                  wallet_import)
 from decentra_network.wallet.print_wallets import print_wallets
