+++ tests/unit_tests/test_wallet.py
@@ -16,15 +16,15 @@ from decentra_network.lib.encryption import decrypt
 from decentra_network.lib.settings_system import (change_wallet, save_settings,
                                                   the_settings)
 from decentra_network.wallet.delete_current_wallet import delete_current_wallet
-from decentra_network.wallet.ellipticcurve.get_saved_wallet import \
+from decentra_network.wallet.get_saved_wallet import \
     get_saved_wallet
 from decentra_network.wallet.ellipticcurve.privateKey import PrivateKey
 from decentra_network.wallet.ellipticcurve.publicKey import PublicKey
-from decentra_network.wallet.ellipticcurve.save_wallet_list import \
+from decentra_network.wallet.save_wallet_list import \
     save_wallet_list
-from decentra_network.wallet.ellipticcurve.wallet_create import wallet_create
-from decentra_network.wallet.ellipticcurve.wallet_delete import wallet_delete
-from decentra_network.wallet.ellipticcurve.wallet_import import (Address,
+from decentra_network.wallet.wallet_create import wallet_create
+from decentra_network.wallet.wallet_delete import wallet_delete
+from decentra_network.wallet.wallet_import import (Address,
                                                                  wallet_import,
                                                                  wallet_import_all)
 from decentra_network.wallet.print_wallets import print_wallets
